requirements.txt
django
django-environ
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-filter